requirements.txt
numpy
pandas
pykrx
yfinance
matplotlib
seaborn
tqdm
requests
python-dotenv
requests
python-dotenv
PyJWT

# 이 프로젝트에서 사용하는 나머지 패키지들이 있으면 아래에 추가
